pyproject.toml
[project]
name = "stealth-browser-mcp"
version = "0.1.0"
description = "Stealth browser MCP server using Patchright"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0",
    "patchright>=1.0",
    "trafilatura>=2.0",
    "readability-lxml>=0.8",
    "html2text>=2024.0",
]

[project.optional-dependencies]
tor = ["stem>=1.8"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stealth_browser"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
